5th Commit
Here is my second update for today.  After sitting in on Ben's
Lunch & Learn, I was able to successfully craft and style a
game board to my liking.  Additionally, I was able to successfully hide
the desired elements based off of certain actions.  However, my new
issue is that I can't bring back the game board once I've hidden it
so that's my next task to work on.  I also decided to go with an 80's
Arcade-theme for the CSS and spent a little time on that today.  The
styling is nowhere near where I'd like it, but I am getting closer to
having the main qualifications down.