6th Commit
Here's my first commit for today.  I was able to resolve the issue of
my game board not generating after the 'new game' button was clicked,
it turns out I had nested the parent elements incorrectly which is why
when I requested it, it couldn't locate it.  I've also adjusted my
jQuery to start including some of my comments/thought process just to
keep better track of all of the moving parts, and have begun the process
of making another 'new game' button to start over once the first game
is completed.