Second half of 10/19 commit
Here is my second commit for 10/19.  I was able to successfully solve
my previous issue, and was also able to create the 'count games' button
too.  I believe the final remaining task is to create the event to
notify when 3 X's or 3 O's create a win! (Plus, CSS) I also added in
// Comments to explain my thought process in a little more detail.