Final commit pre-presentation
I worked very heavily on this today.  I was able to get the majority
of all buttons functioning and was able to log in/log out successfully.
Additionally, I also adjusted my CSS.  The one thing I believe I have
left at this point is the index of the boxes clicked.