Final Commit for Tic-Tac-Toe
Attached is my final commit for my first project.  After the inital
submission, I went back and edited a number of things.  The first thing
I took care of was getting my PATCH API to run successfully.  After that
I had altered my gameboard to be buttons instead of a form, that way it
gave it a few more possibilities to edit in terms of jQuery and
functionality on the users end.  While doing that, I was able to clear
up the issue of new games rotating between X and O (now it starts at X),
and I was able to have it generate a winner message.  It also locks the
screen once a win/tie is detected and it no longer lets you doubleclick
the same box multiple times to cycle between X and O.

The final bug I encountered that I am still working on resolving is now
my "start over" new game button locks the screen entirely once pushed -
it didn't do that until I changed them over to a button.  I am still
researching into why this is happening.  The CSS is still not where I
want it to be, but that's also considered a work in progress.